+++ src/api/get_object.rs
@@ -405,7 +405,8 @@ impl<'a> GetObjectFluentBuilder<'a> {
             accept_ranges: response.get_header("accept-ranges").map(String::from),
             delete_marker: response
                 .get_header("x-amz-delete-marker")
-                .and_then(|s| s.parse::<bool>().ok()),
+                .map(crate::xml::parse_xml_bool)
+                .transpose()?,
             replication_status: response
                 .get_header("x-amz-replication-status")
                 .map(String::from),
@@ -425,7 +426,8 @@ impl<'a> GetObjectFluentBuilder<'a> {
                 .map(String::from),
             bucket_key_enabled: response
                 .get_header("x-amz-server-side-encryption-bucket-key-enabled")
-                .and_then(|s| s.parse::<bool>().ok()),
+                .map(crate::xml::parse_xml_bool)
+                .transpose()?,
             request_charged: response
                 .get_header("x-amz-request-charged")
                 .map(String::from),